move ColorPicker into material/components
